Move apiserver to a package
